cabint page was deconstruct to components
