+++ src/components/forCabinetPage/maps/index.tsx
@@ -1,0 +1,14 @@
+import React,{Fragment} from 'react'
+import style from './maps.module.scss'
+
+const UniversityMaps = ()=>{
+    return(
+        <Fragment>
+        <div className={style.mainstyle}>
+            <p>Карти університетів</p>
+        </div>
+        </Fragment>
+    )
+}
+
+export default UniversityMaps;
