+++ src/components/forCabinetPage/maps/maps.module.scss
@@ -1,0 +1,10 @@
+@import '../../styles.scss';
+
+.mainstyle{
+    width: 100%;
+    height: 23.75rem;
+    background-color: #C4C4C4;
+    font-size:$font_30px;
+    display: flex;
+    justify-content:center;
+}
