+++ src/components/forCabinetPage/options/mainOptions/option1/option1.module.scss
@@ -1,7 +1,6 @@
 @import '../../../../styles.scss';
 
 .mainstyle{
-    height: 34.38rem;
     width: inherit;
     height: 100%;
     background:#e6e6e6;
