+++ src/components/index.tsx
@@ -8,6 +8,10 @@ export { default as BannerLower } from './bannerLower';
 export { default as Partners } from './partners';
 export { default as Footer } from './footer/footer';
 //For cabinet page
+//work space
+export { default as UserWorksSpace } from './forCabinetPage/userWorksSpace/index'
+//maps
+export { default as UniversityMaps} from './forCabinetPage/maps/index'
 // main options
 export { default as Option1MainMenu } from './forCabinetPage/options/mainOptions/option1/index'
 export { default as Option2MainMenu } from './forCabinetPage/options/mainOptions/option2/index'
@@ -20,4 +24,4 @@ export { default as Option2UserMenu } from './forCabinetPage/options/userOptions
 export { default as Option3UserMenu } from './forCabinetPage/options/userOptions/option3/index'
 export { default as Option4UserMenu } from './forCabinetPage/options/userOptions/option4/index'
 export { default as Option5UserMenu } from './forCabinetPage/options/userOptions/option5/index'
-//maps
+
