+++ src/pages/graduateCabinetPage/graduateCabinetPage.module.scss
@@ -1,182 +1,4 @@
-
 .page{
     height: 1008px;
-    .mainWorkingFiled{
-        height: 700px;
-        background-color: #C4C4C4;
-        position: relative;
-        display: flex;
-        justify-content: space-between;
-        .mainMenuBar{
-            width: 45px;
-            height: 700px;   
-            background-color: gray;
-            border:solid 1px black;
-            z-index: 0;
-            position: absolute;
-            //left:20px;
-            cursor: pointer;
-            .logo{
-                font-size:24px;
-                width: 100%;
-                text-align: center;
-                margin-top:11px;
-            }
-            .homeIcon{
-                margin:0 auto;
-                margin-top:50px;
-                text-align: center;
-                padding-top: 6px;
-                width: 33px;
-            }
-            .userIcon{
-                margin:0 auto;
-                margin-top:15px;
-                text-align: center;
-                padding-top: 6px;
-                width: 33px;
-            }            
-            
-        }
-        
-        .mainFiled{ 
-            //border:solid 1px red;
-            margin-left:45px;
-            width: 100%;
-            //margin-left: auto;
-            //margin-right: auto;
-            //position: absolute;
-            //left:20px;
-            z-index: 0;
-            overflow:auto;
-        }
-    }
-    .univMap{
-        //margin-top:55px;
-        height: 380px;
-        background-color: #C4C4C4;
-        text-align: center;
-    }
-}
-/*
-.mainMenuBar .homeIcon:hover + .mainMenu{
-    width: 310px;
-    transition: 0.4s;
-}
-*/
-
-.mainMenu{
-    //width: 310px;
-    width: 0px;
-    height: 700px;
-    left:45px;
-    background-color: gray;
-    position: absolute;
-    z-index: 1;
-    transition: 0.4s;
-    /*
-    &:hover{
-        width: 310px;  
-    }
-    */
-    .userFiled{
-        margin-top:20px;
-        width: 100%;
-        display: flex;
-        align-items: center;
-        overflow-x:hidden;
-        .icon{
-           width: 55px;
-           height: 55px;
-           border-radius:50%;
-           background-color: white;
-           margin-left: 20px;
-        }
-        p{
-            margin-left: 20px;
-        }
-    }
-    .optionList{
-        margin-top:20px;
-        ul{
-            //list-style-type: none;
-            //width: 0px;
-            width: 100%;
-            overflow-x:hidden;
-            li{
-                //width: 100%;
-                height: 30px;
-                overflow:hidden;
-                padding: 7px 18px;
-                border:solid 1px black;
-                cursor:pointer;
-                &:hover{
-                    background-color: white;
-                }
-            }
-        }
-    }
 }
 
-.userMenu{
-    //width: 310px;
-    width: 0px;
-    height: 700px;
-    left:45px;
-    background-color: gray;
-    position: absolute;
-    z-index: 1;
-    transition: 0.4s;
-    .userFiled{
-        margin-top:20px;
-        width: 100%;
-        display: flex;
-        align-items: center;
-        overflow-x:hidden;
-        .icon{
-           width: 55px;
-           height: 55px;
-           border-radius:50%;
-           background-color: white;
-           margin-left: 20px;
-        }
-        p{
-            margin-left: 20px;
-        }
-    }
-    .optionList{
-        margin-top:20px;
-        ul{
-            //list-style-type: none;
-            //width: 0px;
-            width: 100%;
-            overflow-x:hidden;
-            li{
-                //width: 100%;
-                height: 30px;
-                overflow:hidden;
-                padding: 7px 18px;
-                border:solid 1px black;
-                cursor:pointer;
-                &:hover{
-                    background-color: white;
-                }
-            }
-        }
-    }
-
-}
-
-// classes for manipulation on objects
-.whiteBackground{
-    background-color:white;
-}
-.hoverEffectIcons{
-    &:hover{
-        background-color: #BDBBBB;
-    }
-}
-.mainMenuWidth{
-    width: 310px;
-    transition: 0.4s;
-}
